+++ pkg/inst/NEWS.md
@@ -1,11 +1,13 @@
 # Version 3.0.13 [Unreleased]
-* Parallel support added to `hybridModel()`. This can be controlled by setting `parallel = TRUE` (the default) and setting `num.cores`. By default this is not enabled since the performance improvement typically only occurs when fitting `auto.arima` and `tbats` on long series.
+* Parallel support added to `hybridModel()`. This can be controlled by setting `parallel = TRUE` and setting `num.cores`. By default this is not enabled since the performance improvement typically only occurs when fitting `auto.arima` and `tbats` models on long series with large frequency (e.g. `taylor`).
 * Added `z.args` for the `snaive()` model.
 * The `tbats()` and `snaive()` models now respect and use `lambda` when passed in `t.args` and `z.args`.
+* Refactored code to remove code duplication and cleaned up the hybridModel file by moving code into helper and generic files.
+* Weights and cross validation for the `snaive` model are now handled correctly
 
 # Version 2.2.12 [2018-05-04]
 * Added `PI.combination` argument to `forecast.hybridModel()`. The default behavior is to follow the existing methodology of using the most extreme prediction intervals from the component models. When `"mean"` is passed instead, a simple (unweighted) average of the component prediction intervals is used instead.
-* The theta model included in an ensemble can nowhandle seasonality with frequency >= 24.
+* The theta model included in an ensemble can now handle seasonality with frequency >= 24.
 * The ets model can now be included for hourly data.
 * The "reshape2" package is no longer imported.